add updating paper stupplier on company merge add paceholder detection to find_operators turn off some degbugging add supplier to employee paper details

+++ perl/openprint/administrator_managerial.pm
@@ -515,11 +515,17 @@ sub company_profiles {
 				$_->save({'invoicee_id'=>$Company->id()});
 			} # end foreach 
 			foreach my $Payment ( openprint::Payment->find('payor_id'=>$param{'merge_company_id'}) ) {
-				$Payment->save({'payor_id'=>$Company->id()}) if $Payment->payor_id() == $Company->id();
+				$Payment->save({'payor_id'=>$Company->id()});
+ #if $Payment->payor_id() == $Company->id();
 			} # end foreach  Payment
 			foreach my $Payment ( openprint::Payment->find('recipient_id'=>$param{'merge_company_id'}) ) {
-				$Payment->save({'recipient_id'=>$Company->id()}) if $_->recipient_id() == $Company->id();
+				$Payment->save({'recipient_id'=>$Company->id()});
+# if $Payment->recipient_id() == $Company->id();
 			} # end foreach  Payment
+			foreach my $Stock ( openprint::Paper->find(supplier_id=>$param{merge_company_id}) ) {
+				$Stock->save({supplier_id=>$Company->id()});
+ if $_->supplier_id() == $Company->id();
+			} # end foreach  Stock
 			new openprint::Company( $param{'merge_company_id'} )->delete();
 			sql::end_transaction( $dbh, $ac );
 		} # end if
